Refactor make trace conversion more explicit
And move them to the telemetry package as an extension trait
